Fix sign-extension bug in breakpoint instruction poking.
We were using the wrong type for the original instruction, so poking it back
often set the upper word to all-bits-on, causing a later crash.

Orabug: 18674244

+++ libproc/Pcontrol.c
@@ -1857,7 +1857,7 @@ static int
 bkpt_handle_post_singlestep(struct ps_prochandle *P, bkpt_t *bkpt)
 {
 	int state = PS_RUN;
-	int orig_insn;
+	unsigned long orig_insn;
 
 	if (bkpt->after_singlestep) {
 		bkpt_handler_t *notifier = dt_list_next(&bkpt->bkpt_notifiers);
